+++ mintapi/constants.py
@@ -1,0 +1,23 @@
+JSON_FORMAT = "json"
+CSV_FORMAT = "csv"
+
+ACCOUNT_KEY = "Account"
+BUDGET_KEY = "Budget"
+CATEGORY_KEY = "Category"
+INVESTMENT_KEY = "Investment"
+TRANSACTION_KEY = "Transaction"
+NET_WORTH_KEY = "Net_Worth"
+CREDIT_SCORE_KEY = "Credit_Score"
+CREDIT_REPORT_KEY = "Credit_Report"
+
+MINT_ROOT_URL = "https://mint.intuit.com"
+MINT_BETA_ROOT_URL = "https://beta.mint.intuit.com"
+MINT_CREDIT_URL = "https://credit.finance.intuit.com"
+
+JSON_HEADER = {"accept": "application/json"}
+
+MFA_VIA_SOFT_TOKEN = "soft-token"
+MFA_VIA_AUTHENTICATOR = "authenticator"
+MFA_VIA_EMAIL = "email"
+MFA_VIA_SMS = "sms"
+MFA_METHOD_LABEL = "mfa_method"