+++ tests/test_driver.py
@@ -5,7 +5,7 @@ import json
 import unittest
 import requests
 import tempfile
-
+from mintapi import constants
 from unittest.mock import patch, DEFAULT
 
 
@@ -272,8 +272,10 @@ class MintApiTests(unittest.TestCase):
     ):
         test_exception = Exception()
         mock_sign_in.side_effect = test_exception
-        mintapi.Mint("test", "test")
+        with self.assertRaises(Exception) as context:
+            mintapi.Mint("test", "test")
         mock_logger.exception.assert_called_with(test_exception)
+        self.assertTrue("Could not sign in to Mint" in str(context.exception))
 
     @patch.multiple(
         mintapi.Mint,
@@ -286,27 +288,27 @@ class MintApiTests(unittest.TestCase):
     )
     def test_exclude_credit_details(self, **_):
         mint = mintapi.Mint()
-        credit_report = mint.get_credit_report(
+        credit_report = mint.get_credit_report_data(
             limit=2, details=True, exclude_inquiries=True
         )
         self.assertFalse("inquiries" in credit_report)
-        credit_report = mint.get_credit_report(
+        credit_report = mint.get_credit_report_data(
             limit=2, details=True, exclude_inquiries=False
         )
         self.assertTrue("inquiries" in credit_report)
-        credit_report = mint.get_credit_report(
+        credit_report = mint.get_credit_report_data(
             limit=2, details=True, exclude_accounts=True
         )
         self.assertFalse("accounts" in credit_report)
-        credit_report = mint.get_credit_report(
+        credit_report = mint.get_credit_report_data(
             limit=2, details=True, exclude_accounts=False
         )
         self.assertTrue("accounts" in credit_report)
-        credit_report = mint.get_credit_report(
+        credit_report = mint.get_credit_report_data(
             limit=2, details=True, exclude_utilization=True
         )
         self.assertFalse("utilization" in credit_report)
-        credit_report = mint.get_credit_report(
+        credit_report = mint.get_credit_report_data(
             limit=2, details=True, exclude_utilization=False
         )
         self.assertTrue("utilization" in credit_report)
@@ -318,63 +320,68 @@ class MintApiTests(unittest.TestCase):
         self.assertEqual(arguments.transactions, True)
         config_file.close()
 
-    @patch.object(mintapi.Mint, "_Mint__call_accounts_endpoint")
-    def test_get_investment_data_new(self, mock_call_accounts_endpoint):
+    @patch.object(mintapi.Mint, "_Mint__call_mint_endpoint")
+    def test_get_account_data(self, mock_call_accounts_endpoint):
         mock_call_accounts_endpoint.return_value = accounts_example
         account_data = mintapi.Mint().get_account_data()[0]
         self.assertFalse("metaData" in account_data)
         self.assertTrue("createdDate" in account_data)
         self.assertTrue("lastUpdatedDate" in account_data)
 
-    @patch.object(mintapi.Mint, "_Mint__call_transactions_endpoint")
+    @patch.object(mintapi.Mint, "_Mint__call_mint_endpoint")
     def test_get_transaction_data(self, mock_call_transactions_endpoint):
         mock_call_transactions_endpoint.return_value = transactions_example
         transaction_data = mintapi.Mint().get_transaction_data()[0]
         self.assertFalse("metaData" in transaction_data)
+        self.assertFalse("createdDate" in transaction_data)
         self.assertTrue("lastUpdatedDate" in transaction_data)
         self.assertTrue("parentId" in transaction_data["category"])
         self.assertTrue("parentName" in transaction_data["category"])
 
-    @patch.object(mintapi.Mint, "_Mint__call_investments_endpoint")
-    def test_get_investment_data_new(self, mock_call_investments_endpoint):
+    @patch.object(mintapi.Mint, "_Mint__call_mint_endpoint")
+    def test_get_investment_data(self, mock_call_investments_endpoint):
         mock_call_investments_endpoint.return_value = investments_example
         investment_data = mintapi.Mint().get_investment_data()[0]
         self.assertFalse("metaData" in investment_data)
+        self.assertFalse("createdDate" in investment_data)
         self.assertTrue("lastUpdatedDate" in investment_data)
 
-    @patch.object(mintapi.Mint, "_Mint__call_budgets_endpoint")
+    @patch.object(mintapi.Mint, "_Mint__call_mint_endpoint")
     def test_get_budgets(self, mock_call_budgets_endpoint):
         mock_call_budgets_endpoint.return_value = budgets_example
-        budgets = mintapi.Mint().get_budgets()[0]
+        budgets = mintapi.Mint().get_budget_data()[0]
         self.assertFalse("metaData" in budgets)
+        self.assertTrue("createdDate" in budgets)
         self.assertTrue("lastUpdatedDate" in budgets)
 
     def test_format_filename(self):
         config_file = write_transactions_file()
         arguments = parse_arguments_file(config_file)
-        filename = mintapi.cli.format_filename(arguments)
-        self.assertEqual(filename, "transactions.csv")
+        type = constants.TRANSACTION_KEY.lower()
+        filename = mintapi.cli.format_filename(arguments, type)
+        self.assertEqual(filename, "current_{}.csv".format(type))
 
         config_file = write_accounts_file()
         arguments = parse_arguments_file(config_file)
-        filename = mintapi.cli.format_filename(arguments)
-        self.assertEqual(filename, "accounts.json")
+        type = constants.ACCOUNT_KEY.lower()
+        filename = mintapi.cli.format_filename(arguments, type)
+        self.assertEqual(filename, "current_{}.json".format(type))
 
         config_file = write_investments_file()
         arguments = parse_arguments_file(config_file)
-        filename = mintapi.cli.format_filename(arguments)
+        filename = mintapi.cli.format_filename(arguments, None)
         self.assertEqual(filename, None)
 
 
 def write_transactions_file():
     config_file = tempfile.NamedTemporaryFile(mode="wt")
-    config_file.write("transactions\nformat=csv\nfilename=transactions")
+    config_file.write("transactions\nformat=csv\nfilename=current")
     return config_file
 
 
 def write_accounts_file():
     config_file = tempfile.NamedTemporaryFile(mode="wt")
-    config_file.write("accounts\nformat=json\nfilename=accounts")
+    config_file.write("accounts\nformat=json\nfilename=current")
     return config_file