+++ CHANGELOG.md
@@ -1,20 +1,39 @@
-2.0 (Pending)
----
+--> NOTE! The 1.x release series only supports and contain fixes for the previous Mint UI.
+    New feature work occurs in the 2.x release series and only supports the new Mint UI / endpoints.
+    Please see the README for more information.
 
+2.1
+---
+- Add `--beta` switch to force using the new Mint endpoints (#470)
+- Fix for credit score data (#472)
+
+2.0
+---
+- Dynamic Multifactor Authentication Flow (#392)
 - Add Data Format Option to CLI: `--format` (#432)
-- Update Accounts Endpoint to meet new Mint requirements (#430) 
-- Update Transactions Endpoint to meet new Mint requirements (#429) 
+- Update Accounts Endpoint to meet new Mint requirements (#430)
+- Update Transactions Endpoint to meet new Mint requirements (#429)
 - Update Budgets Endpoint to meet new Mint requirements (#425)
 - Support removed for hiding duplicate transactions (#427)
 - Fetch the correct API Key to use with Mint API Requests (#420, #423)
 - Removed `get_token` functionality, which is incompatible with the new Mint UI (#421)
 - Update the name of the Account Refresh Class (#415)
-- Update the overview page url (#414) 
+- Update the overview page url (#414)
 
-BREAKING CHANGES:
-- Because of the new Mint UI and the switchover to different API Endpoints, the data structure associated with each function may be different.  Please verify that the data you are expecting against the new output of each endpoint.
-- To help support the new CLI option for data format (`--format`), we have eliminated the need to specify a file extension in `--filename`.  Be aware that if you do not specify `--format=csv`, then the extension\format will be `json`, which means that any filename you specify will include the extension of `.json`.
+:warning: BREAKING CHANGES :warning:
+- As of version 2.0, mintapi only supports the new Mint.com UI and endpoints. Please see the README to learn which version of mintapi you need.
+- `mfa_method` is only required if your login flow presents you with the option to select which Multifactor Authentication Method you wish to use, typically as a result of your account configured to accept different methods.  We now automatically detect the type of MFA requested by Mint, based on the prompts on the screen.  Because of this, mintapi now supports the use case where multiple MFA prompts appear.
+- Due to the new Mint UI and the switchover to different API Endpoints, the data structure associated with each function may be different.  Please verify that the data you are expecting against the new output of each endpoint.
+- To implement consistency and to call attention to the fact that some input and output of methods has changed, some API methods have been renamed.  With the exception of bills, all data fetch methods now follow the same sort of pattern: `get_<data type>_data`.  For example, instead of `get_accounts`, we now have `get_account_data`.
+- To help support the new CLI option for data format (`--format`), we have eliminated the need to specify a file extension in `--filename`.  Be aware that if you do not specify `--format=csv`, then the extension/format will be `json`, and any filename will automatically include the extension of `.json`. In addition, the CLI now supports receiving multiple types of data in one call to mintapi. See the "Multi-Data Support" section of the README for more information.
 
+1.66
+---
+- handle "Sign in without a password next time" page (#436)
+
+1.65
+---
+- fix for selenium dependency version (#412, #413, #416)
 
 1.64
 ---